abstracted fixture
