+++ configs/server_config.yaml
@@ -4,16 +4,16 @@ persona = sam_pov
 human = basic
 
 [model]
-model = gpt-4
-model_endpoint = https://api.openai.com/v1
-model_endpoint_type = openai
+model = ehartford/dolphin-2.5-mixtral-8x7b
+model_endpoint = https://api.memgpt.ai
+model_endpoint_type = vllm
 context_window = 8192
 
 [embedding]
-embedding_endpoint_type = openai
-embedding_endpoint = https://api.openai.com/v1
-embedding_model = text-embedding-ada-002
-embedding_dim = 1536
+embedding_endpoint_type = hugging-face
+embedding_endpoint = https://embeddings.memgpt.ai
+embedding_model = BAAI/bge-large-en-v1.5
+embedding_dim = 1024
 embedding_chunk_size = 300
 
 [archival_storage]
